feat: Update Prisma schema to cascade delete for cuentaBancariaOrigen
The Prisma schema has been updated to change the onDelete behavior for the relation "cuentaBancariaOrigen" in the "Operacion" model. Previously, it was set to "NoAction", but now it has been changed to "Cascade". This change ensures that when a bank account is deleted, all associated operations will also be deleted.